deleted soap message object
Delete soap message object and move the functionality to soap body
